Base Class Creation
-- made the base class:  DataLogger.
-- updated PSQLDataLogger to inherit from DataLogger.
-- removed unused methods
-- some typo fixes
